+++ modules/features/discover/src/main/java/au/com/shiftyjelly/pocketcasts/discover/view/DiscoverAdapter.kt
@@ -26,6 +26,7 @@ import au.com.shiftyjelly.pocketcasts.discover.databinding.RowCategoryAdBinding
 import au.com.shiftyjelly.pocketcasts.discover.databinding.RowCategoryPillsBinding
 import au.com.shiftyjelly.pocketcasts.discover.databinding.RowChangeRegionBinding
 import au.com.shiftyjelly.pocketcasts.discover.databinding.RowCollectionListBinding
+import au.com.shiftyjelly.pocketcasts.discover.databinding.RowCollectionListDeprecatedBinding
 import au.com.shiftyjelly.pocketcasts.discover.databinding.RowErrorBinding
 import au.com.shiftyjelly.pocketcasts.discover.databinding.RowMostPopularPodcastsBinding
 import au.com.shiftyjelly.pocketcasts.discover.databinding.RowPodcastLargeListBinding
@@ -65,6 +66,8 @@ import au.com.shiftyjelly.pocketcasts.ui.theme.Theme
 import au.com.shiftyjelly.pocketcasts.utils.Optional
 import au.com.shiftyjelly.pocketcasts.utils.extensions.dpToPx
 import au.com.shiftyjelly.pocketcasts.utils.extensions.toLocalizedFormatPattern
+import au.com.shiftyjelly.pocketcasts.utils.featureflag.Feature.GUEST_LISTS_NETWORK_HIGHLIGHTS_REDESIGN
+import au.com.shiftyjelly.pocketcasts.utils.featureflag.FeatureFlag
 import au.com.shiftyjelly.pocketcasts.views.extensions.show
 import au.com.shiftyjelly.pocketcasts.views.extensions.showIf
 import coil.imageLoader
@@ -80,6 +83,7 @@ import io.reactivex.rxkotlin.subscribeBy
 import io.reactivex.rxkotlin.zipWith
 import io.reactivex.schedulers.Schedulers
 import java.util.Locale
+import kotlin.math.ceil
 import kotlin.math.min
 import kotlinx.coroutines.MainScope
 import kotlinx.coroutines.launch
@@ -440,7 +444,47 @@ internal class DiscoverAdapter(
 
     class SinglePodcastViewHolder(val binding: RowSinglePodcastBinding) : NetworkLoadableViewHolder(binding.root)
     class SingleEpisodeViewHolder(val binding: RowSingleEpisodeBinding) : NetworkLoadableViewHolder(binding.root)
-    class CollectionListViewHolder(val binding: RowCollectionListBinding) : NetworkLoadableViewHolder(binding.root)
+    class CollectionListDeprecatedViewHolder(val binding: RowCollectionListDeprecatedBinding) : NetworkLoadableViewHolder(binding.root)
+
+    inner class CollectionListViewHolder(val binding: RowCollectionListBinding) : NetworkLoadableViewHolder(binding.root), ShowAllRow {
+        val adapter = CollectionListRowAdapter(listener::onPodcastClicked, listener::onPodcastSubscribe, analyticsTracker)
+
+        override val showAllButton: TextView
+            get() = binding.btnShowAll
+
+        private val linearLayoutManager =
+            LinearLayoutManager(itemView.context, RecyclerView.HORIZONTAL, false).apply {
+                initialPrefetchItemCount = 2
+            }
+
+        init {
+            recyclerView?.layoutManager = linearLayoutManager
+            recyclerView?.itemAnimator = null
+            val snapHelper = HorizontalPeekSnapHelper(0.dpToPx(itemView.context))
+            snapHelper.attachToRecyclerView(recyclerView)
+            snapHelper.onSnapPositionChanged = { position ->
+                binding.pageIndicatorView.position = position
+                val row = getItem(bindingAdapterPosition) as? DiscoverRow
+                row?.let {
+                    analyticsTracker.track(
+                        AnalyticsEvent.DISCOVER_COLLECTION_LIST_PAGE_CHANGED,
+                        mapOf(CURRENT_PAGE to position, TOTAL_PAGES to adapter.itemCount, LIST_ID_KEY to it.inferredId()),
+                    )
+                }
+            }
+
+            recyclerView?.adapter = adapter
+
+            adapter.showLoadingList()
+        }
+
+        override fun onRestoreInstanceState(state: Parcelable?) {
+            super.onRestoreInstanceState(state)
+            recyclerView?.post {
+                binding.pageIndicatorView.position = linearLayoutManager.findFirstVisibleItemPosition()
+            }
+        }
+    }
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
         val inflater = LayoutInflater.from(parent.context)
@@ -464,6 +508,7 @@ internal class DiscoverAdapter(
             R.layout.row_category_ad -> CategoryAdViewHolder(RowCategoryAdBinding.inflate(inflater, parent, false))
             R.layout.row_single_podcast -> SinglePodcastViewHolder(RowSinglePodcastBinding.inflate(inflater, parent, false))
             R.layout.row_single_episode -> SingleEpisodeViewHolder(RowSingleEpisodeBinding.inflate(inflater, parent, false))
+            R.layout.row_collection_list_deprecated -> CollectionListDeprecatedViewHolder(RowCollectionListDeprecatedBinding.inflate(inflater, parent, false))
             R.layout.row_collection_list -> CollectionListViewHolder(RowCollectionListBinding.inflate(inflater, parent, false))
             else -> ErrorViewHolder(RowErrorBinding.inflate(inflater, parent, false))
         }
@@ -484,13 +529,15 @@ internal class DiscoverAdapter(
                         is DisplayStyle.LargeList -> R.layout.row_podcast_large_list
                         is DisplayStyle.SmallList -> R.layout.row_podcast_small_list
                         is DisplayStyle.SinglePodcast -> R.layout.row_single_podcast
-                        is DisplayStyle.CollectionList -> R.layout.row_collection_list
+                        is DisplayStyle.CollectionList ->
+                            if (FeatureFlag.isEnabled(GUEST_LISTS_NETWORK_HIGHLIGHTS_REDESIGN)) R.layout.row_collection_list else R.layout.row_collection_list_deprecated
                         else -> R.layout.row_error
                     }
                 } else if (row.type is ListType.EpisodeList) {
                     return when (row.displayStyle) {
                         is DisplayStyle.SingleEpisode -> R.layout.row_single_episode
-                        is DisplayStyle.CollectionList -> R.layout.row_collection_list
+                        is DisplayStyle.CollectionList ->
+                            if (FeatureFlag.isEnabled(GUEST_LISTS_NETWORK_HIGHLIGHTS_REDESIGN)) R.layout.row_collection_list else R.layout.row_collection_list_deprecated
                         else -> R.layout.row_error
                     }
                 } else if (row.type is ListType.Categories && row.displayStyle is DisplayStyle.Pills) {
@@ -720,7 +767,7 @@ internal class DiscoverAdapter(
                     )
                 }
 
-                is CollectionListViewHolder -> {
+                is CollectionListDeprecatedViewHolder -> {
                     holder.loadFlowable(
                         loadPodcastList(row.source),
                         onNext = {
@@ -765,6 +812,21 @@ internal class DiscoverAdapter(
                             onRestoreInstanceState(holder)
 
                             row.listUuid?.let { listUuid -> trackListImpression(listUuid) }
+                        },
+                    )
+                }
+                is CollectionListViewHolder -> {
+                    holder.loadFlowable(
+                        loadPodcastList(row.source),
+                        onNext = {
+                            val podcasts = it.podcasts.subList(0, MAX_ROWS_SMALL_LIST.coerceAtMost(it.podcasts.count()))
+                            holder.binding.pageIndicatorView.count = ceil(podcasts.count().toDouble() / CollectionListRowAdapter.CollectionListViewHolder.NUMBER_OF_ROWS_PER_PAGE.toDouble()).toInt()
+
+                            row.listUuid?.let { listUuid -> holder.adapter.setFromListId(listUuid) }
+
+                            holder.binding.lblTitle.text = it.subtitle?.tryToLocalise(resources)
+
+                            holder.adapter.submitPodcastList(podcasts) { onRestoreInstanceState(holder) }
                         },
                     )
                 }