+++ modules/services/repositories/src/main/java/au/com/shiftyjelly/pocketcasts/repositories/shownotes/ShowNotesProcessor.kt
@@ -12,8 +12,6 @@ import au.com.shiftyjelly.pocketcasts.servers.podcast.PodcastCacheService
 import au.com.shiftyjelly.pocketcasts.servers.podcast.ShowNotesChapter
 import au.com.shiftyjelly.pocketcasts.servers.podcast.ShowNotesResponse
 import au.com.shiftyjelly.pocketcasts.servers.podcast.ShowNotesTranscript
-import au.com.shiftyjelly.pocketcasts.utils.featureflag.Feature
-import au.com.shiftyjelly.pocketcasts.utils.featureflag.FeatureFlag
 import javax.inject.Inject
 import kotlin.time.Duration.Companion.seconds
 import kotlinx.coroutines.CoroutineScope
@@ -35,9 +33,7 @@ class ShowNotesProcessor @Inject constructor(
         updateImageUrls(showNotes)
         updateChapters(episodeUuid, showNotes)
         updateChapterFromLink(episodeUuid, showNotes)
-        if (FeatureFlag.isEnabled(Feature.TRANSCRIPTS)) {
-            updateTranscripts(episodeUuid, showNotes, loadTranscriptSource)
-        }
+        updateTranscripts(episodeUuid, showNotes, loadTranscriptSource)
     }
 
     private fun updateImageUrls(showNotes: ShowNotesResponse) = scope.launch {
@@ -86,11 +82,10 @@ class ShowNotesProcessor @Inject constructor(
         showNotes: ShowNotesResponse,
         loadTranscriptSource: LoadTranscriptSource,
     ) = scope.launch {
-        val transcripts = showNotes.podcast?.episodes
-            ?.firstOrNull { it.uuid == episodeUuid }
-            ?.transcripts
-            ?.mapNotNull { it.takeIf { it.url != null && it.type != null }?.toTranscript(episodeUuid) }
-        transcripts?.let { transcriptsManager.updateTranscripts(showNotes.podcast?.uuid.orEmpty(), episodeUuid, it, loadTranscriptSource) }
+        val transcripts = showNotes.findTranscripts(episodeUuid)
+        if (transcripts != null) {
+            transcriptsManager.updateTranscripts(showNotes.podcast?.uuid.orEmpty(), episodeUuid, transcripts, loadTranscriptSource)
+        }
     }
 
     private fun ShowNotesChapter.toDbChapter(index: Int, episodeUuid: String) = if (useInTableOfContents != false) {
@@ -109,9 +104,24 @@ class ShowNotesProcessor @Inject constructor(
     }
 }
 
-fun ShowNotesTranscript.toTranscript(episodeUuid: String) = Transcript(
-    episodeUuid = episodeUuid,
-    url = requireNotNull(url),
-    type = requireNotNull(type),
-    language = language,
-)
+internal fun ShowNotesResponse.findTranscripts(episodeUuid: String): List<Transcript>? {
+    val episode = podcast?.episodes?.firstOrNull { it.uuid == episodeUuid } ?: return null
+    val transcripts = episode.transcripts?.mapNotNull { it.toTranscript(episodeUuid, isGenerated = false) }.orEmpty()
+    val pocketCastsTranscripts = episode.pocketCastsTranscripts?.mapNotNull { it.toTranscript(episodeUuid, isGenerated = true) }.orEmpty()
+    return transcripts + pocketCastsTranscripts
+}
+
+private fun ShowNotesTranscript.toTranscript(
+    episodeUuid: String,
+    isGenerated: Boolean,
+) = if (url != null && type != null) {
+    Transcript(
+        episodeUuid = episodeUuid,
+        url = requireNotNull(url),
+        type = requireNotNull(type),
+        isGenerated = isGenerated,
+        language = language,
+    )
+} else {
+    null
+}