+++ CHANGELOG.md
@@ -3,12 +3,24 @@
 *   New Features
     *   Add smart folders support.
         ([#3720](https://github.com/Automattic/pocket-casts-android/pull/3720))
+    *   Add generated transcripts support.
+        ([#3772](https://github.com/Automattic/pocket-casts-android/pull/3772))
 *   Bug Fixes
+    *   Fixed an issue where the Up Next queue obstructed the bottom navigation bar in some cases.
+        ([#3762](https://github.com/Automattic/pocket-casts-android/pull/3762))
+    *   Fixed an issue where the navigation bar was drawn over the player in some cases.
+        ([#3762](https://github.com/Automattic/pocket-casts-android/pull/3762))
     *   Fixed an issue where the mini player could sometimes hide behind the navigation bar.
         ([#3687](https://github.com/Automattic/pocket-casts-android/pull/3687))
+    *   Fixed an issue where bookmark confirmation button was obstructed by keyboard.
+        ([#3761](https://github.com/Automattic/pocket-casts-android/pull/3761))
 *   Updates
     *   Close episode details screen when archiving an episode. This reverts [#3473](https://github.com/Automattic/pocket-casts-android/pull/3473). 
         ([#3718](https://github.com/Automattic/pocket-casts-android/pull/3718))
+    *   Update widget titles and add descriptions.
+        ([#3769](https://github.com/Automattic/pocket-casts-android/pull/3769))
+    *   Reorder podcasts sorting order based on their popularity. 
+        ([#3733](https://github.com/Automattic/pocket-casts-android/pull/3733))
 
 7.84
 -----
@@ -22,6 +34,8 @@
         ([#3666](https://github.com/Automattic/pocket-casts-android/pull/3666))
     *   Fix screen flickering when tapping on item in the navigation bar.
         ([#3674](https://github.com/Automattic/pocket-casts-android/pull/3674))
+    *   Fix crash with podcast update when offline.
+        ([#3718](https://github.com/Automattic/pocket-casts-android/pull/3718))
 
 7.83
 -----